1. Renames queryRunner() to queryRunnerReducer() 2. Moves IAction to it's own file
