+++ src/app/services/reducers/index.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import { queryRunner } from './query-runner-reducers';
+import { queryRunnerReducer } from './query-runner-reducers';
 
 export default combineReducers({
-  graphResponse: queryRunner,
+  graphResponse: queryRunnerReducer,
 });
