+++ src/app/services/reducers/query-runner-reducers.ts
@@ -1,7 +1,7 @@
-import { IAction } from '../../../types/query-runner';
+import { IAction } from '../../../types/action';
 import { QUERY_GRAPH_ERROR, QUERY_GRAPH_SUCCESS } from '../constants';
 
-export function queryRunner(state = {}, action: IAction) {
+export function queryRunnerReducer(state = {}, action: IAction) {
   switch (action.type) {
     case QUERY_GRAPH_SUCCESS:
       return action.response;
