+++ src/types/action.ts
@@ -1,0 +1,4 @@
+export interface IAction {
+  type: string;
+  response: string;
+}
